Refactors thread generation into a Python package
